Add files via upload
I was thinking about add elements of prime factor in the list and find the largest prime factor in the list. It worked in small range. However, when range was increased it gave me an error or even it worked but it took a lot of times. So I changed like this. I think it is the best way to solve this question.