Add files via upload
This is the solution for Euler project #13. I learned the how to convert input data. Like if you look at my code, data1 was received as list type. So I did not that much code to find the answer. However, I like my first approach because I learned that range of slicing include the end spot. Also, it makes me think the question.