Add files via upload
This is solution for Euler project #14. I spent a lot of time because I did not understand the question correctly at the first time. I thought I had to find the number of chain when the number is 1000000. However, the question was I have to find the number which has maximum chain in the range of 1 to 1000000. Finally, I got the right answer but it still need to fix it because it took much time to calculate the number.